Update the project structure to work with kotlin-master
