+++ gradle-conventions/build.gradle.kts
@@ -10,12 +10,10 @@ plugins {
     alias(libs.plugins.gradle.kotlin.dsl)
 }
 
-// chicken-egg
-apply(from = "src/main/kotlin/compiler-specific-module.gradle.kts")
-
 defaultConventionConfiguration()
 
 dependencies {
+    implementation(project(":common"))
     implementation(":gradle-conventions-settings")
 
     project.whenKotlinLatest {
