+++ gradle-conventions/common/src/main/kotlin/build-util.gradle.kts
@@ -1,0 +1,5 @@
+/*
+ * Copyright 2023-2024 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
+ */
+
+// empty
