+++ gradle-conventions/common/src/main/kotlin/util/JsTarget.kt
