+++ gradle-conventions/common/src/main/kotlin/util/OptionalProperty.kt
