+++ gradle-conventions/common/src/main/kotlin/util/ProjectKotlinConfig.kt
