+++ gradle-conventions/common/src/main/kotlin/util/PublicModule.kt
