+++ gradle-conventions/common/src/main/kotlin/util/PublicationUtils.kt
