+++ gradle-conventions/common/src/main/kotlin/util/StringCase.kt
