+++ gradle-conventions/common/src/main/kotlin/util/atomicfu.kt
