+++ gradle-conventions/common/src/main/kotlin/util/npm.kt
