+++ gradle-conventions/common/src/main/kotlin/util/properties.kt
