+++ gradle-conventions/settings.gradle.kts
@@ -5,10 +5,6 @@
 import util.otherwise
 import util.whenKotlinLatest
 
-/*
- * Copyright 2023-2024 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
- */
-
 rootProject.name = "gradle-conventions"
 
 enableFeaturePreview("TYPESAFE_PROJECT_ACCESSORS")
@@ -23,6 +19,8 @@ plugins {
     id("conventions-repositories")
     id("conventions-version-resolution")
 }
+
+include(":common")
 
 whenKotlinLatest {
     include(":latest-only")
